+++ vsg/rules/port/rule_604.py
@@ -1,6 +1,8 @@
 
 from vsg.rules import token_prefix as Rule
 
+from vsg.rules import utils
+
 from vsg import token
 
 lTokens = []
@@ -41,21 +43,5 @@ class rule_604(Rule):
     def _get_tokens_of_interest(self, oFile):
         lReturn = []
         lToi = oFile.get_interface_elements_between_tokens(token.port_clause.open_parenthesis, token.port_clause.close_parenthesis)
-        lReturn = extract_identifiers_with_mode_of_linkage(lToi)
+        lReturn = utils.extract_identifiers_with_mode_of_linkage(lToi)
         return lReturn
-
-
-def extract_identifiers_with_mode_of_linkage(lToi):
-    lReturn = []
-    for oToi in lToi:
-        if interface_element_mode_is_linkage(oToi):
-            lReturn.append(oToi.extract_tokens(0, 0))
-    return lReturn
-
-
-def interface_element_mode_is_linkage(oToi):
-    lTokens = oToi.get_tokens()
-    for oToken in lTokens:
-        if isinstance(oToken, token.mode.linkage_keyword):
-            return True
-    return False
