+++ vsg/rules/port/rule_606.py
@@ -1,6 +1,8 @@
 
 from vsg.rules import token_suffix as Rule
 
+from vsg.rules import utils
+
 from vsg import token
 
 lTokens = []
@@ -40,21 +42,5 @@ class rule_606(Rule):
     def _get_tokens_of_interest(self, oFile):
         lReturn = []
         lToi = oFile.get_interface_elements_between_tokens(token.port_clause.open_parenthesis, token.port_clause.close_parenthesis)
-        lReturn = extract_identifiers_with_mode_of_out(lToi)
+        lReturn = utils.extract_identifiers_with_mode_of_out(lToi)
         return lReturn
-
-
-def extract_identifiers_with_mode_of_out(lToi):
-    lReturn = []
-    for oToi in lToi:
-        if interface_element_mode_is_output(oToi):
-            lReturn.append(oToi.extract_tokens(0, 0))
-    return lReturn
-
-
-def interface_element_mode_is_output(oToi):
-    lTokens = oToi.get_tokens()
-    for oToken in lTokens:
-        if isinstance(oToken, token.mode.out_keyword):
-            return True
-    return False
