+++ vsg/rules/port/rule_608.py
@@ -1,6 +1,8 @@
 
 from vsg.rules import token_suffix as Rule
 
+from vsg.rules import utils
+
 from vsg import token
 
 lTokens = []
@@ -40,21 +42,5 @@ class rule_608(Rule):
     def _get_tokens_of_interest(self, oFile):
         lReturn = []
         lToi = oFile.get_interface_elements_between_tokens(token.port_clause.open_parenthesis, token.port_clause.close_parenthesis)
-        lReturn = extract_identifiers_with_mode_of_buffer(lToi)
+        lReturn = utils.extract_identifiers_with_mode_of_buffer(lToi)
         return lReturn
-
-
-def extract_identifiers_with_mode_of_buffer(lToi):
-    lReturn = []
-    for oToi in lToi:
-        if interface_element_mode_is_buffer(oToi):
-            lReturn.append(oToi.extract_tokens(0, 0))
-    return lReturn
-
-
-def interface_element_mode_is_buffer(oToi):
-    lTokens = oToi.get_tokens()
-    for oToken in lTokens:
-        if isinstance(oToken, token.mode.buffer_keyword):
-            return True
-    return False
