+++ vsg/vhdlFile/extract/tokens.py
@@ -45,6 +45,12 @@ class New():
     def get_end_index(self):
         return self.iEndIndex
 
+    def token_type_exists(self, oTokenType):
+        for oToken in self.lTokens:
+            if isinstance(oToken, oTokenType):
+                return True
+        return False
+
 
 def calculate_end_index(iStartIndex, lTokens):
     try:
